Renaming id to uuid in tests
RISDEV-4560
